Turned into a fun puzzle experience
It was fun trying to improve the accuracy of the "prompt"(going into the embedding model) from the blog, but accuracy remained the same with higher precision but lower recall. It should be noted that the data is too skewed and scores don't necessarily correlate with the review opinion e.g. negative reviews could be complementing the product while saying that it was not what they had anticipated or that they don't like anything among this kind of products.